+++ hztb-rest-api/src/main/java/com/svs/hztb/api/sm/model/refresh/OpinionCountData.java
@@ -1,27 +1,32 @@
 package com.svs.hztb.api.sm.model.refresh;
 
 public class OpinionCountData {
-	private int userId;
+	private Long userId;
 	private int givenCount;
 	private int pendingCount;
-	public int getUserId() {
+
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(int userId) {
+
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
+
 	public int getGivenCount() {
 		return givenCount;
 	}
+
 	public void setGivenCount(int givenCount) {
 		this.givenCount = givenCount;
 	}
+
 	public int getPendingCount() {
 		return pendingCount;
 	}
+
 	public void setPendingCount(int pendingCount) {
 		this.pendingCount = pendingCount;
 	}
-	
-	
+
 }
