+++ hztb-rest-api/src/main/java/com/svs/hztb/api/sm/model/user/UserData.java
@@ -1,28 +1,32 @@
 package com.svs.hztb.api.sm.model.user;
 
 public class UserData {
-	private Integer userId;
-	private String firstName="";
-	private String lastname="";
-	public Integer getUserId() {
+	private Long userId;
+	private String firstName = "";
+	private String lastname = "";
+
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(Integer userId) {
+
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
+
 	public String getFirstName() {
 		return firstName;
 	}
+
 	public void setFirstName(String firstName) {
 		this.firstName = firstName;
 	}
+
 	public String getLastname() {
 		return lastname;
 	}
+
 	public void setLastname(String lastname) {
 		this.lastname = lastname;
 	}
-	
-	
-	
+
 }
