+++ hztb-servicemanager/src/main/java/com/svs/hztb/entity/OpinionResponseEntity.java
@@ -6,8 +6,6 @@ import java.util.Date;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.NamedQuery;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
@@ -19,7 +17,7 @@ import javax.persistence.TemporalType;
  */
 @Entity
 @Table(name = "opinion_response", schema = "ebdb")
- @NamedQuery(name="OpinionResponse.findAll", query="SELECT o FROM OpinionResponseEntity o")
+@NamedQuery(name = "OpinionResponse.findAll", query = "SELECT o FROM OpinionResponseEntity o")
 public class OpinionResponseEntity implements Serializable {
 	private static final long serialVersionUID = 1L;
 
@@ -28,13 +26,11 @@ public class OpinionResponseEntity implements Serializable {
 	private int responseId;
 
 	@Column(name = "responder_user_id")
-	private int responderUserId;
-
+	private Long responderUserId;
 
 	@Column(name = "response_txt")
 	private String responseTxt;
 
-	
 	@Column(name = "opinion_id")
 	private int opinionId;
 
@@ -42,9 +38,9 @@ public class OpinionResponseEntity implements Serializable {
 	private String opinionResponseType;
 
 	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name="response_time", insertable = false)
+	@Column(name = "response_time", insertable = false)
 	private Date responseTime;
-	
+
 	public OpinionResponseEntity() {
 	}
 
@@ -56,11 +52,11 @@ public class OpinionResponseEntity implements Serializable {
 		this.responseId = responseId;
 	}
 
-	public int getResponderUserId() {
+	public Long getResponderUserId() {
 		return this.responderUserId;
 	}
 
-	public void setResponderUserId(int responderUserId) {
+	public void setResponderUserId(Long responderUserId) {
 		this.responderUserId = responderUserId;
 	}
 
@@ -100,5 +96,4 @@ public class OpinionResponseEntity implements Serializable {
 		this.responseTime = responseTime;
 	}
 
-	
 }
